+++ nodemon.json
@@ -1,6 +1,6 @@
 {
-    "watch": ["src"],
+    "watch": ["server"],
     "ext": ".ts,.js",
     "ignore": [],
-    "exec": "ts-node ./src/Server.ts"
+    "exec": "ts-node ./server/Server.ts"
 }
