+++ package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start:dev": "nodemon",
-    "start":"ts-node ./src/Server.ts"
+    "start":"ts-node ./server/Server.ts"
 
   },
   "keywords": [],
