+++ server/api/model/user/IUser.ts
@@ -1,0 +1,9 @@
+/**
+ * User interface.
+ */
+export interface IUser {
+    firstName: string,
+    lastName: string,
+    uid: string,
+    lastSeen: number
+}
