--- src/api/model/user/IUser.ts
@@ -1,6 +1,0 @@
-export interface IUser {
-    firstName: string,
-    lastName: string,
-    uid: string,
-    lastSeen: number
-}
