--- src/api/view/html/public/index.html
@@ -1,28 +1,0 @@
-<!DOCTYPE html>
-<!--[if lt IE 7]>      <html class="no-js lt-ie9 lt-ie8 lt-ie7"> <![endif]-->
-<!--[if IE 7]>         <html class="no-js lt-ie9 lt-ie8"> <![endif]-->
-<!--[if IE 8]>         <html class="no-js lt-ie9"> <![endif]-->
-<!--[if gt IE 8]>      <html class="no-js"> <!--<![endif]-->
-<html>
-
-<head>
-    <meta charset="utf-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <title></title>
-    <meta name="description" content="">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link rel="stylesheet" href="">
-    <script type="module" src="auth.js"></script>
-</head>
-
-<body>
-    <h1>Home Page</h1>
-    <script src="" async defer></script>
-    <div class="row">
-        <input id="email-label"/>
-        <input id="password-label"/>
-        <button id="login-btn">Login</button>        
-    </div>
-</body>
-
-</html>
