--- src/authentication/IAuthenticator.ts
@@ -1,5 +1,0 @@
-import { NextFunction, Request, Response } from "express";
-
-export interface IAuthenticator {
-    authenticate(req: Request, res: Response, next: NextFunction): void
-}
