--- src/types/express/types.d.ts
@@ -1,9 +1,0 @@
-import { Request } from "express-serve-static-core"; 
-
-declare module "express-serve-static-core" {
-    export interface Request {
-        uid?: string;
-    }
-}
-
-export { };
